feat: create plugin and add access handler for plugin config validation
This feature adds the initial validation of plugin schema with
associated plugin configuration using the access handler to read the
JSON body of a proxy request and perform MetaSubSchema validation along
with plugin subschema validation. Once the schema is validated the
configuration is validated against the supplied schema by gathering the
defaults for the config based on the schema definition and then
inserting the plugin for final validation.